add hooks rules and refactor
