+++ .eslintrc
@@ -1,5 +1,6 @@
 {
   "extends": ["airbnb", "prettier"],
+  "plugins": ["react-hooks"],
   "env": {
     "jest": true
   },
@@ -11,6 +12,8 @@
       "error",
       { "extensions": [".js", ".jsx"] }
     ],
-    "import/prefer-default-export": false
+    "import/prefer-default-export": false,
+    "react-hooks/rules-of-hooks": "error",
+    "react-hooks/exhaustive-deps": "warn"
   }
 }
