+++ package.json
@@ -68,6 +68,7 @@
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-react": "^7.11.0",
+    "eslint-plugin-react-hooks": "^1.5.0",
     "jest": "^24.3.1",
     "jest-dom": "^2.1.0",
     "react": "^16.7.0-alpha.0",
